+++ packages/localize-tools/testdata/build-transform-xliff/goldens/tsout/en/foo.js
@@ -52,3 +52,5 @@ export class MyElement2 extends LitElement {
     return html`<p>Hello <b>World</b>! (${getLocale()})</p>`;
   }
 }
+// Escaped markup characters should remain escaped
+html`&lt;Hello<b>&lt;World &amp; Friends&gt;</b>!&gt;`;