+++ packages/localize-tools/CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+## Fixed
+
+- Escaped `<`, `>`, and `&` characters in HTML text content are now preserved
+  when generating runtime & transform mode output. Previously they sometimes
+  were emitted unescaped, generating invalid markup.
 
 ## [0.3.5] - 2021-07-14